refactor(auth): User component with signing out and user status
